==> src/01_data_extraction.ipynb <==
# 1. Data Extraction
---

The data will be extracted from Kaggle using its web Api service.

from kaggle.api.kaggle_api_extended import KaggleApi
api = KaggleApi()
api.authenticate()

api.dataset_download_files('filipechavesdemacedo/counter-strike-competitive-data', path='../data/raw/', unzip=True)

---
End of data extraction
