==> src/02_database_creation.ipynb <==
# 2. DATABASE CREATION
---

The database used for this proyect will be created using sqlite as main tool. 

From the given tables, a Query will be performed to join them into a new single table available for work and analysis.

### 2.1. Creating database

# Import libraries
import sqlite3 as sq
import pandas as pd
import os
import platform

# Create the .db file
con = sq.connect("SPECIFY HERE WHERE YOU'D WANT TO CREATE THE DB")

# Prepare the data from the .csv file
match_players = pd.read_csv('../data/raw/match_players.csv')
match_results = pd.read_csv('../data/raw/match_results.csv')
players_stats = pd.read_csv('../data/raw/players_stats.csv')

# Insert the dataframe into the .db file

match_players.to_sql('match_players', con, index=False, if_exists='replace')
match_results.to_sql('match_results', con, index=False, if_exists='replace')
players_stats.to_sql('players_stats', con, index=False, if_exists='replace')
con.commit()


### 2.2. Creating dataset from Query

***Writting query***

# WRITE THE QUERY (REMOVING COLUMNS WITHOUT MEANINGFUL USE AS NAMES, LINKS, ETC.)
query = """ 
SELECT adr,
assists, 
deaths, 
fkdiff, 
hs, 
kdratio, 
kills,
rating,
players.match_id, 
player_id, 
team_1_id, 
team_1_name, 
team_1_score, 
team_2_id, 
team_2_name, 
team_2_score,
team_name,
data_unix,
map, 
hour, 
day, 
week, 
month, 
year, 
weekday
FROM match_players as players
LEFT JOIN match_results as results ON players.match_id = results.match_id
"""

***Executing query***

# EXECUTE QUERY TO CREATE A NEW DATASET
total_data = pd.read_sql_query(query,con)

***Saving dataset in folder***

save_dir = '../data/interim/player_match_data.csv'

if platform.system() == 'Windows' and (not os.path.exists('../data/interim')):
    os.mkdir('../data/interim')

total_data.to_csv(save_dir,index=False)

***Closing database connection***

# Close the cursor and connection when done
con.close()

---
End of database creation
