==> src/05_feature_engineering.ipynb <==
# 5 Feature Engineering
---

The initial dataset was constructed from the player’s viewpoint. However, to align with the machine learning model’s requirements, the focus need to be shifted to the team’s perspective. 

Consequently, The dataset will be reshaped  by merging and selectively deleting variables, emphasizing the team-centric information. This transformation will enable the model to accurately predict match outcomes.

### 5.0. Libraries and dataset

# Libraries
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt

# Dataset
total_data = pd.read_csv('../data/interim/player_match_data.csv')
total_data.shape

### 5.1. Creating new variables

***Adding the target variables and extra ones***

    # Variable 1: player_team_id 

# 'player_team_id' function
def find_most_common_team_number(group):
        all_team_ids = pd.concat([group['team_1_id'], group['team_2_id']])
        team_numbers = all_team_ids.dropna().astype(int)
        most_common_team_number = np.argmax(np.bincount(team_numbers))
        return most_common_team_number

# Applying function to dataset
total_data['player_team_id'] = total_data.groupby('team_name').apply(find_most_common_team_number).reindex(total_data['team_name']).values

# ----------------------------------------------------------------------------------------------------------------------------

    # Variable 2: winning_team

# 'winning_team' function
def get_winning_team(team_1_score, team_2_score):
        if team_1_score == team_2_score: return 0
        elif team_1_score > team_2_score: return 1
        else: return 2

# Aapplying function to dataset
total_data['winning_team'] = total_data.apply(lambda row: get_winning_team(row['team_1_score'], row['team_2_score']), axis=1)

# ----------------------------------------------------------------------------------------------------------------------------

    # Variable 3: winning_team_id 

# Creating column called 'winning_team_id' (doesn't need function)
total_data['winning_team_id'] = np.where(total_data['winning_team'] == 1, total_data['team_1_id'],
                                np.where(total_data['winning_team'] == 2, total_data['team_2_id'], 0))

# ----------------------------------------------------------------------------------------------------------------------------

    # Variable 4: player_has_won 

# Creating column called 'player_has_won' (doesn't need function)
total_data['player_has_won'] = np.where(total_data['winning_team_id'] == total_data['player_team_id'], 1, 0)

# ----------------------------------------------------------------------------------------------------------------------------

# Seleccting target variables

target_1 = 'winning_team'
target_2 = 'player_has_won'

# Show small overview of the dataset
total_data.head(3)


***Creating a dataset with all the teams id and names***

To help the dataset transformation, a table with all the team names and their IDs will be created using the main dataset as reference.

unique_teams_1 = total_data[['team_1_id', 'team_1_name']].rename(columns={'team_1_id': 'team_id', 'team_1_name': 'team_name'})
unique_teams_2 = total_data[['team_2_id', 'team_2_name']].rename(columns={'team_2_id': 'team_id', 'team_2_name': 'team_name'})
unique_teams = pd.concat([unique_teams_1, unique_teams_2]).drop_duplicates().reset_index(drop=True)
unique_teams.info()

# SAVE THE DATASET
unique_teams.to_csv('../data/interim/team_data.csv', index=False)

***Create a team dataframe: Group all matchs by MatchID and TeamID***

A sub dataframe will be created with the team information of each match. This will provide a dataframe that has two rows for each match (1 per team). It will have all the players IDs, the mean of adr and kdratio and the sum of the rest of the variables. This will provide us the **teams' performance** in every match played. 

total_data = pd.merge(total_data, unique_teams, left_on='team_name', right_on='team_name', how='left')

# Make invalid teams searchable
total_data = total_data.drop('team_name', axis=1)
total_data['team_id'] = total_data['team_id'].fillna(-1).astype(int)

# Keep only the records with a valid team
total_data = total_data[total_data['team_id'] != -1]

total_data.head()

# Parse the non-numerical fields
total_data["adr"] = pd.to_numeric(total_data["adr"], errors="coerce")
total_data["kdratio"] = (
    pd.to_numeric(total_data["kdratio"].str.rstrip("%"), errors="coerce") / 100.0
)

total_data['adr'] = total_data.groupby('player_id')['adr'].transform(lambda x: x.fillna(x[x.notna()].mean()))
total_data['kdratio'] = total_data.groupby('player_id')['kdratio'].transform(lambda x: x.fillna(x[x.notna()].mean()))

# Deleting rows with null values
null_rows = total_data.loc[total_data.isnull().any(axis=1)]

total_data = total_data.dropna()


***Deleting irrelevant variables***

- The group ID and name variables have been removed because it is not possible to use them for the machine learning model.
- Time group variables were deleted due to their poor performance with both target variables.
- Although the target 2 allowed us to discover the correlation between variables, the target needed for the ML model is the targe 1 -winning team-.


# Drop unused columns for simplicity
total_data.drop(['hour', 'day', 'week', 'month', 'weekday', 'year', 'rating', 'team_1_name', 'team_2_name'], axis=1, inplace=True)
total_data.drop(['team_1_score', 'team_2_score', 'winning_team_id', 'player_has_won'], axis=1, inplace=True)
total_data.head()

# Group by MatchID and TeamID, then calculate average stats and collect PlayerIDs
team_stats = total_data.groupby(['match_id', 'team_id']).agg({
    'player_id': lambda x: frozenset(x),
    'adr': 'mean',
    'assists': 'sum',
    'deaths': 'sum',
    'fkdiff': 'sum',
    'hs': 'sum',
    'kdratio': 'mean',
    'kills': 'sum',
}).reset_index()

team_stats.head()



# Add a new column with the size of frozenset
team_stats['team_size'] = team_stats['player_id'].apply(len)

# Sort the DataFrame by the size of frozenset
team_stats = team_stats.sort_values(by='team_size')

match_ids = team_stats[team_stats['team_size'] != 5]['match_id'].to_list()

team_stats = team_stats[~team_stats['match_id'].isin(match_ids)]

# # Drop the temporary 'team_size' column if needed
team_stats.drop(columns=['team_size'], inplace=True)

team_stats.rename(columns={
    'player_id':'members',
    'adr': 'avg_adr',
    'assists': 'sum_assists',
    'deaths': 'sum_deaths',
    'fkdiff': 'sum_fkdiffs',
    'hs': 'sum_hs',
    'kdratio': 'mean_kdratio',
    'kills': 'sum_kills',
    }, inplace=True)

team_stats.head()

***Merge the teams dataframe with the main dataframe***

The merging will provide the main dataset with the statistics related to the team. Now, every row will have the information on the player and the team views. However, the player information is not needed anymore, so everything but the players' id will be deleted. Finally, the dataframe will be grouped by match ID and team ID, giving two rows per match, one for each team.

# Merge with the original DataFrame to get PlayerIDs
merged_df = pd.merge(total_data, team_stats, how='left', left_on=['match_id', 'team_1_id'], right_on=['match_id', 'team_id'])
merged_df = pd.merge(merged_df, team_stats, how='left', left_on=['match_id', 'team_2_id'], right_on=['match_id', 'team_id'], suffixes=('_team_1', '_team_2'))


merged_df.head(10)


# Drop unnecessary columns and duplicates
merged_df.drop(
    [
        'adr', 
        'assists',
        'deaths',
        'fkdiff',
        'hs',
        'kdratio',
        'kills',
        'player_id',
        'player_team_id',
        'team_id_x',
        'team_id_y',
        'team_id',
        'team_1_id',
        'team_2_id'
        ], 
    axis=1,
    inplace=True
    )


merged_df.sort_values(by='match_id')
merged_df.drop_duplicates(ignore_index=True, inplace=True)
merged_df.head()

***Drop irrelevant columns and reorder the final dataframe***

The final step will consist in rordering the columns and dropping the remain duplicates, nan values and irrelevant columns of the dataset.

# Drop irrelevant columns
total_data = merged_df
total_data.drop([
    'match_id',
    'data_unix'
], axis=1, inplace=True)

# Reorder the dataframe
order = [
    'members_team_1',
    'members_team_2',
    'map',  
    'avg_adr_team_1',
    'sum_assists_team_1', 
    'sum_deaths_team_1', 
    'sum_fkdiffs_team_1',
    'sum_hs_team_1', 
    'mean_kdratio_team_1', 
    'sum_kills_team_1',     
    'avg_adr_team_2', 
    'sum_assists_team_2',
    'sum_deaths_team_2', 
    'sum_fkdiffs_team_2', 
    'sum_hs_team_2',
    'mean_kdratio_team_2', 
    'sum_kills_team_2',
    'winning_team'
]

total_data = total_data.reindex(columns=order)

total_data.dropna(inplace=True)
total_data.head()

***Save the dataset***

total_data.to_csv('../data/interim/clean_match_data.csv',index=False)

---

## Conclusions

The new dataset includes one row per match played. This row has the map type, all the average and sum of the team statistics and also includes the players IDs ordered by team. The target is the winning team. This new dataset has 84040 rows, 18 columns and no null values.
