==> src/06_model_training.ipynb <==
# 6. Model Training

----

This notebook will be perform a machine learning model to predict the outcome of a cs-go match. The model selected is an Artificial Neural Network, due to its capacity to learn from context.

### 6.0. Import libraries and dataset

# Libraries
import pandas as pd
import numpy as np
import seaborn as sns
import matplotlib.pyplot as plt
import os
import platform

# Dataset
total_data = pd.read_csv('../data/interim/clean_match_data.csv')
total_data['members_team_1'] = total_data['members_team_1'].apply(eval)
total_data['members_team_2'] = total_data['members_team_2'].apply(eval)
total_data.head()

### 6.1. Organize the data for the model

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from tensorflow.keras.models import Model, Sequential
from tensorflow.keras.layers import Input, Dense, Concatenate, Embedding, Flatten, concatenate, Dropout

X = total_data[[
    'members_team_1',
    'members_team_2',
    'map',  
    'avg_adr_team_1',
    'sum_assists_team_1', 
    'sum_deaths_team_1', 
    'sum_fkdiffs_team_1',
    'sum_hs_team_1', 
    'mean_kdratio_team_1', 
    'sum_kills_team_1',     
    'avg_adr_team_2', 
    'sum_assists_team_2',
    'sum_deaths_team_2', 
    'sum_fkdiffs_team_2', 
    'sum_hs_team_2',
    'mean_kdratio_team_2', 
    'sum_kills_team_2'
]]

y = total_data['winning_team']

numerical_features = [
    'avg_adr_team_1',
    'sum_assists_team_1', 
    'sum_deaths_team_1', 
    'sum_fkdiffs_team_1',
    'sum_hs_team_1', 
    'mean_kdratio_team_1', 
    'sum_kills_team_1',     
    'avg_adr_team_2', 
    'sum_assists_team_2',
    'sum_deaths_team_2', 
    'sum_fkdiffs_team_2', 
    'sum_hs_team_2',
    'mean_kdratio_team_2', 
    'sum_kills_team_2']

# Split the data into training and testing sets
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=69, stratify=y)

***Transform the data into usable format***


from pickle import dump

X_train_frozenset_1 = X_train[['members_team_1']]
X_test_frozenset_1 = X_test[['members_team_1']]

X_train_frozenset_2 = X_train[['members_team_2']]
X_test_frozenset_2 = X_test[['members_team_2']]

# Standardize numerical features
scaler = StandardScaler()
X_train[numerical_features] = scaler.fit_transform(X_train[numerical_features])
X_test[numerical_features] = scaler.transform(X_test[numerical_features])

# One-hot encode categorical features
encoder_categorical = OneHotEncoder(sparse_output=False, drop='first')
X_train['map'] = encoder_categorical.fit_transform(X_train[['map']])
X_test['map'] = encoder_categorical.transform(X_test[['map']])

# Determine the number of categories dynamically
num_categories = len(total_data['members_team_1'].unique()) + len(total_data['members_team_2'].unique())

# Use a hash function to map frozensets to integers
hash_function = lambda x: hash(x)
X_train['members_team_1'] = X_train_frozenset_1['members_team_1'].apply(hash_function)
X_test['members_team_1'] = X_test_frozenset_1['members_team_1'].apply(hash_function)

X_train['members_team_2'] = X_train_frozenset_2['members_team_2'].apply(hash_function)
X_test['members_team_2'] = X_test_frozenset_2['members_team_2'].apply(hash_function)

# Standardize numerical features
X_train[['members_team_1', 'members_team_2']] = scaler.fit_transform(X_train[['members_team_1', 'members_team_2']])
X_test[['members_team_1', 'members_team_2']] = scaler.transform(X_test[['members_team_1', 'members_team_2']])




***Save the scalers used***

if platform.system() == 'Windows' and (not os.path.exists('../models/scalers')):
    os.mkdir('../models/scalers')

    # Save scalers
with open("../models/scalers/onehot_scaler.pkl", "wb") as f:
    dump(encoder_categorical, f)

with open("../models/scalers/standard_scaler.pkl", "wb") as f:
    dump(scaler, f)

### 6.2. Create the neural network

# Input layers for each branch
input1 = Input(shape=(2,))
input2 = Input(shape=(len(X_train.columns)-2,))

dense_layer1 = Dense(64, activation='relu')(input1)
flatten1 = Flatten()(dense_layer1)

# Dense layer for the second input
dense_layer2 = Dense(32, activation='relu')(input2)
flatten2 = Flatten()(dense_layer2)

# Add more hidden layers
dense_layer3 = Dense(64, activation='relu')(flatten1)
dense_layer4 = Dense(64, activation='relu')(flatten2)

dropout1 = Dropout(0.5)(dense_layer3)
dropout2 = Dropout(0.5)(dense_layer4)

concatenated = concatenate([dropout1, dropout2])

# Additional layers for further processing
dense1 = Dense(128, activation='relu')(concatenated)
output_layer = Dense(1, activation='sigmoid')(dense1)

# Create the model
model = Model(inputs=[input1, input2], outputs=output_layer)

# Compile the model
model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])

# Display the model summary
model.summary()

### 6.3. Train the model

# Train the model
model.fit([X_train[['members_team_1','members_team_2']], pd.concat([X_train['map'], X_train[numerical_features]], axis=1)], y_train, epochs=20, batch_size=32, validation_split=0.1)

### 6.4. Evaluate the model

# Evaluate the model on the test set
loss = model.evaluate([X_test[['members_team_1','members_team_2']], pd.concat([X_test['map'], X_test[numerical_features]], axis=1)], y_test)
print(f'Mean Squared Error on Test Set: {loss}')

### 6.5. Save the model

model.save("../models/csgo_match_results_predictor-beta1.h5")

### 6.6. Testing the model

temp = X_test.iloc[0]

len(temp)

temp = X_test.iloc[0]

temp[2:len(temp)] = 0.0

input1_data = np.array([[-0.025245, 0.959970]])  
input2_data = np.array([temp.iloc[2:].values])  

predictions = model.predict([input1_data, input2_data])

# predictions 

predictions

total_data.head()

# Dataset for players
player_data = pd.read_csv('../data/raw/match_players.csv')

player_data = player_data[['player_id', 'player_nick']].drop_duplicates()

# SAVE THE DATASET

if platform.system() == 'Windows' and (not os.path.exists('../data/extras')):
    os.mkdir('../data/extras')

player_data.to_csv('../data/extras/player_data.csv', index=False)
